% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-version_comparison.R
\name{fig03_coverage_panel}
\alias{fig03_coverage_panel}
\title{Figure 3A -- overall resource coverage}
\usage{
fig03_coverage_panel(data, width_mm = 180L)
}
\arguments{
\item{data}{Harmonized MPI rows.}

\item{width_mm}{Numeric target width.}
}
\value{
A ggplot object.
}
\description{
Figure 3A -- overall resource coverage
}
