% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-version_comparison.R
\name{fig03_metabolite_class_panel}
\alias{fig03_metabolite_class_panel}
\title{Figure 3B -- metabolite-class coverage}
\usage{
fig03_metabolite_class_panel(data, width_mm = 180L, top_n = 10L)
}
\arguments{
\item{data}{Harmonized MPI rows.}

\item{width_mm}{Numeric target width.}

\item{top_n}{Integer number of classes to keep.}
}
\value{
A ggplot object.
}
\description{
Figure 3B -- metabolite-class coverage
}
