% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-version_comparison.R
\name{fig03_source_relationship_panel}
\alias{fig03_source_relationship_panel}
\title{Figure 3E -- source by relationship type}
\usage{
fig03_source_relationship_panel(data, width_mm = 180L)
}
\arguments{
\item{data}{Harmonized MPI rows.}

\item{width_mm}{Numeric target width.}
}
\value{
A ggplot object.
}
\description{
Figure 3E -- source by relationship type
}
