% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-palettes.R
\name{palette_rwth}
\alias{palette_rwth}
\title{The rwth extended palette}
\usage{
palette_rwth()
}
\value{
Named character vector of hex codes.
}
\description{
Parses \code{inst/extdata/palettes/rwth.gpl} into a named hex
vector. Used for categorical needs that exceed the lead palette
(FR-021).
}
\examples{
head(palette_rwth())

}
