% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\description{
Re-export of magrittr's \verb{\\\%>\\\%} so internal callers in this package
resolve it at install / load time.
}
\keyword{internal}
