% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-db_content.R
\name{plot_annotation_classes_by_resource}
\alias{plot_annotation_classes_by_resource}
\title{Panel E — annotation classes by resource}
\usage{
plot_annotation_classes_by_resource(data, width_mm = 89L)
}
\arguments{
\item{data}{Tibble from \code{\link{annotation_classes_by_resource}}.}

\item{width_mm}{Numeric.}
}
\value{
A ggplot object.
}
\description{
Panel E — annotation classes by resource
}
