% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-category_colours.R
\name{category_colour}
\alias{category_colour}
\title{Get the colour for a (category, value) pair}
\usage{
category_colour(category, value)
}
\arguments{
\item{category}{Character scalar: one of the registered category
names (e.g. \code{"interaction_types"}, \code{"resources"}).}

\item{value}{Character: one or more category values to look up.}
}
\value{
Character vector of hex codes, same length as \code{value}.
}
\description{
Get the colour for a (category, value) pair
}
\examples{
category_colour("interaction_types", c("signaling", "transport"))

}
