% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-db_content.R
\name{interactions_by_resource}
\alias{interactions_by_resource}
\title{Number of interactions per upstream resource}
\usage{
interactions_by_resource(con)
}
\arguments{
\item{con}{A DBI connection.}
}
\value{
A tibble with columns \code{resource} and \code{n_relations}.
}
\description{
Figure 1 Panel C.
}
