% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-db_content.R
\name{ontology_terms_by_ontology}
\alias{ontology_terms_by_ontology}
\title{Number of ontology terms per ontology}
\usage{
ontology_terms_by_ontology(con)
}
\arguments{
\item{con}{A DBI connection.}
}
\value{
A tibble with columns \code{ontology} and \code{n_terms}.
}
\description{
Figure 1 Panel F.
}
