% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-db_content.R
\name{plot_ontology_terms_by_ontology}
\alias{plot_ontology_terms_by_ontology}
\title{Panel F — ontology terms by ontology}
\usage{
plot_ontology_terms_by_ontology(data, width_mm = 89L)
}
\arguments{
\item{data}{Tibble from \code{\link{ontology_terms_by_ontology}}.}

\item{width_mm}{Numeric.}
}
\value{
A ggplot object.
}
\description{
Panel F — ontology terms by ontology
}
